feat: connect backend and refactor
